# backend/swagger.yaml
openapi: 3.0.0
info:
  title: InvoiceGuard API
  description: API для сервиса приема платежей (Middle+ Implementation)
  version: 1.0.0

servers:
  - url: http://localhost:3000/api
    description: Local server

paths:
  /auth/register:
    post:
      summary: Регистрация нового пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: "admin@d9911.org" }
                password: { type: string, example: "d9911" }
      responses:
        201:
          description: Пользователь создан

  /auth/login:
    post:
      summary: Вход в систему
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: "admin@d9911.org" }
                password: { type: string, example: "d9911" }
      responses:
        200:
          description: Успешный вход или требование 2FA

  /auth/2fa/enable:
    post:
      summary: Генерация секрета 2FA (TOTP)
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: Возвращает секрет и QR-код в формате DataURL

  /auth/2fa/verify:
    post:
      summary: Подтверждение и включение 2FA
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, example: "123456", description: "6-значный код из Google Authenticator" }
      responses:
        200:
          description: 2FA успешно включен

  /auth/2fa/login-verify:
    post:
      summary: Проверка 2FA при входе
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, example: "123456" }
      responses:
        200:
          description: Возвращает финальный JWT токен

  /invoice:
    post:
      summary: Создание счета (invoice)
      tags: [Invoices]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, currency, merchantId]
              properties:
                amount: { type: integer, example: 10000, description: "Сумма в копейках/центах" }
                currency: { type: string, example: "USD" }
                merchantId: { type: string, example: "merchant_123" }
      responses:
        201:
          description: Счет создан

  /invoice/{id}:
    get:
      summary: Получение статуса счета
      tags: [Invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Информация о счете
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'

  /webhook:
    post:
      summary: Прием статуса оплаты (Webhook)
      tags: [Webhooks]
      description: |
        Требует заголовки безопасности:
        - X-Signature: HMAC-SHA256(raw_body, merchant_secret)
        - X-Timestamp: Unix timestamp (seconds)
        - X-Nonce: Уникальная строка
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invoiceId, status]
              properties:
                invoiceId: { type: string, format: uuid }
                status: { type: string, enum: [paid, failed] }
      responses:
        200:
          description: Статус обновлен

  /health:
    get:
      summary: Проверка работоспособности
      tags: [System]
      responses:
        200:
          description: Статус системы

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Invoice:
      type: object
      properties:
        invoiceId: { type: string }
        status: { type: string, enum: [pending, paid, failed] }
        amount: { type: integer }
        currency: { type: string }
        fee: { type: integer }
        amountToReceive: { type: integer }